Renamed search-new to search, removed old search page, added pagination
