+++ src/app/search/_components/result.module.css
